Remove multiple consecutive blank lines by default
This ensures the generated Markdown does not violate the following markdownlint rule:

  MD012 - Multiple consecutive blank lines

  https://github.com/DavidAnson/markdownlint/blob/v0.23.1/doc/Rules.md#md012

Add a property to the "Config" object that allows this behavior to be disabled.

Update existing test that violates this markdownlint rule with the new default behavior (i.e. when Config.RemoveMultipleConsecutiveBlankLines == true) so that it continues to pass (without changing the expected results).

Add a new test that verifies multiple consecutive blank lines are removed as expected with the default configuration value (i.e. true).