Remove consecutive blank lines at beginning/end
Note: This change would require updates to numerous test results (if enabled for those tests). To avoid updating the "verified" output -- and, more importantly, to continue treating them as "Markdown fragments" instead of "complete Markdown documents" -- I opted to modify the failing tests to disable the option to remove consecutive blank lines.

This preserves the ability to identify where whitespace and line endings are being generated and/or trimmed by the various "converter" classes.