Refactor: Implement generic "Convert" method in ConverterBase
- Most of the element converters implement a similar pattern in their Convert methods:

  1. Convert the child nodes from HTML to Markdown by calling ConverterBase.TreatChildren

  2. Take the content from the previous step and wrap it inside some variant of "{prefix}{content}{suffix}" -- where "{prefix}" and "{suffix}" would typically be Environment.NewLine (at least for block elements like <aside>, <h1>, <h2>, ..., <hr>, and <p>). For <code> inline elements, the prefix and suffix are both a string containing a single backquote ("`").

  3. Sometimes -- depending on the type of element being converted -- the Markdown generated by converting child nodes is trimmed.

This refactoring introduces new intermediate classes for the various block and inline elements. Block element converters will typically use Environment.NewLine for the prefix and suffix, while inline elements will typically not have a prefix or suffix (or, rather more accurately speaking, will use string.Empty for the prefix and suffix).

Note that block element converters can easily override the "{prefix}" as is the case with <blockquote> in order to insert a blank line above the content in the Markdown.

In general, element converters should override the following methods as necessary:

- GetMarkdownContent
- GetMarkdownPrefix
- GetMarkdownSuffix

However, in some cases, such when processing headings, the base Convert method may need to be overridden (see H.Convert for an example and more detail on why this is necessary).