Fix malformed content in HTML lists
There was previously code in the "P" converter to handle the scenario where <p> elements are found directly inside lists (<ol> and <ul>). It greatly simplifies things going forward if this logic is removed from the low-level element converter and instead handled during the "cleanup" phase prior to Markdown conversion.