Move code from BlockElementConverter to DefaultMarkdownFormatter
This refactoring makes the code much easier to test (due to the growing number of formatting rules added to BlockElementConverter.GetMarkdownContent).

The "rules" for formatting Markdown are determined by the IMarkdownFormatter.GetFormattingRules method -- which now has its own set of tests.

Most of the code previously in BlockElementConverter.GetMarkdownContent moves to DefaultMarkdownFormatter.GetFormattingRules.

Once the rules are determined, the actual formatting process is much simpler (and therefore remains in the BlockElementConverter.GetMarkdownContent method).

Note the refactoring also requires moving code from ConverterBase (specifically the GetMarkdownPrefixForListItem method) -- which, in turn, requires passing the Config dependency when creating new instances of DefaultMarkdownFormatter.